python/image_processing/requirements.txt
Wand==0.6.1
Pillow==7.1.2
CairoSVG==2.4.2
